Concluindo validações Modal livro, editora, categoria e autor.

+++ src/main/java/com/dac/ecommerce/livros/controller/AutorController.java
@@ -1,5 +1,7 @@
 package com.dac.ecommerce.livros.controller;
 
+import java.util.List;
+
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -13,7 +15,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import com.dac.ecommerce.livros.dto.DTOAutor;
 import com.dac.ecommerce.livros.model.livro.Autor;
+import com.dac.ecommerce.livros.model.livro.Livro;
 import com.dac.ecommerce.livros.services.AutorService;
+import com.dac.ecommerce.livros.services.LivroService;
 
 @Controller
 @RequestMapping("/autor")
@@ -22,23 +26,27 @@ public class AutorController {
 	@Autowired
 	private AutorService autorService;
 	
+	@Autowired
+	private LivroService livroService;
+
 	@RequestMapping("/menu-autor")
-	public String menu(Model model) {
-		model.addAttribute("autores",autorService.todosAutores());
-		model.addAttribute("dtoAutor",new DTOAutor());
+	public String menu(Model model, Model autores) {
+		model.addAttribute("autores", autorService.todosAutores());
+		model.addAttribute("dtoAutor", new DTOAutor());
 		return "/autor/menu-autor";
 	}
-	
+
 	@GetMapping("/cadastrar-autor")
 	public String form(Model model) {
-		model.addAttribute("dtoAutor",new DTOAutor());
+		model.addAttribute("dtoAutor", new DTOAutor());
 		return "/autor/cadastrar-autor";
 	}
 
 	@PostMapping("/salvar")
-	public String salvar(@Valid @ModelAttribute("dtoAutor")DTOAutor dtoAutor, BindingResult bindingResult, RedirectAttributes atts) {
-		
-		if(!bindingResult.hasErrors()) {
+	public String salvar(@Valid @ModelAttribute("dtoAutor") DTOAutor dtoAutor, BindingResult bindingResult,
+			RedirectAttributes atts) {
+
+		if (!bindingResult.hasErrors()) {
 			Autor autor = dtoAutor.toAutor();
 			autorService.salvar(autor);
 			return "redirect:/autor/menu-autor";
@@ -48,22 +56,38 @@ public class AutorController {
 	}
 
 	@GetMapping("/deletar/{id}")
-	public String deletar(@PathVariable("id") Long id) {
-		autorService.remove(id);
-		return "redirect:/autor/menu-autor";
+	public String deletar(@PathVariable("id") Long id) throws Exception {
+		
+		List<Livro> livros = livroService.recuperarTodosOsLivros();
+		Autor autor = autorService.recuperarAutor(id);
+		boolean result = false;
+		if(!livros.isEmpty()) {
+			for (int i = 0; i < livros.size(); i++) {
+				if(livros.get(i).getAutores().contains(autor)) {
+					result = true;
+					break;
+				}
+			}
+		}
+		if(result == false) {
+			autorService.remove(id);
+			return "redirect:/autor/menu-autor";
+		}
+		return "/autor/erro-excluir-autor";
 	}
 
 	@PostMapping("/alterar-autor")
-	public String alterar(@Valid @ModelAttribute("dtoAutor") DTOAutor dtoAutor, BindingResult bindingResult, RedirectAttributes atts) {
+	public String alterar(@Valid @ModelAttribute("dtoAutor") DTOAutor dtoAutor, BindingResult bindingResult,
+			RedirectAttributes atts, Model model) {
 		
-		if(!bindingResult.hasErrors()) {
+		if (!bindingResult.hasErrors()) {
 			Autor autor = dtoAutor.toAutor();
-			System.out.println(dtoAutor);
 			autorService.editarAutor(autor, autor.getId());
 			return "redirect:/autor/menu-autor";
 		}
+		model.addAttribute("autores", autorService.todosAutores());
 		atts.addAttribute("hasErrors", true);
-		return "redirect:/autor/menu-autor";
+		return "autor/menu-autor";
 	}
 
 }
